added abstract SectionRule class
