+++ src/main/kotlin/com/github/darderion/mundaneassignmentpolice/checker/rule/section/SectionSizeRule.kt
@@ -1,21 +1,18 @@
 package com.github.darderion.mundaneassignmentpolice.checker.rule.section
 
-import com.github.darderion.mundaneassignmentpolice.checker.RuleViolation
 import com.github.darderion.mundaneassignmentpolice.checker.RuleViolationType
-import com.github.darderion.mundaneassignmentpolice.checker.rule.Rule
 import com.github.darderion.mundaneassignmentpolice.pdfdocument.PDFArea
 import com.github.darderion.mundaneassignmentpolice.pdfdocument.PDFDocument
-import com.github.darderion.mundaneassignmentpolice.pdfdocument.PDFRegion.Companion.NOWHERE
 import com.github.darderion.mundaneassignmentpolice.pdfdocument.text.Section
 
 class SectionSizeRule(
     name: String,
     type: RuleViolationType,
-    val sectionName: String,
-    val pageLimit: Int,
-    val percentageLimit: Int
-): Rule(NOWHERE.except(PDFArea.SECTION), name, type) {
-    private fun isViolated(section: Section, document: PDFDocument): Boolean {
+    sectionName: String,
+    private val pageLimit: Int,
+    private val percentageLimit: Int
+): SectionRule(name, type, sectionName) {
+    override fun isViolated(section: Section, document: PDFDocument): Boolean {
         val text = document.text
         val sections = document.areas!!.sections
 
@@ -42,15 +39,5 @@ class SectionSizeRule(
         val totalPages = text.last().page + 1
         val sectionSize = lastSectionPage - firstSectionPage + 1
         return sectionSize > pageLimit || sectionSize.toDouble() / totalPages * 100 > percentageLimit
-    }
-
-    override fun process(document: PDFDocument): List<RuleViolation> {
-        val section = document.areas!!.sections.firstOrNull {
-            it.title.contains(sectionName, true)
-        } ?: return emptyList()
-
-        if (!isViolated(section, document)) return emptyList()
-
-        return listOf(RuleViolation(listOf(document.text[section.titleIndex]), name, type))
     }
 }
